HSA-52: Add role based Authorization for a house
Added HouseRolesGuard and @HouseRoles decorator for role based authorization:

The decorator works with the HouseRolesGuard. Add these two together to protect an API endpoint with defined roles.

Make sure some where (either in parent or in the current controller function) there is a param named 'house_id'. The role uses this route_param to check if the user is a member of the house, then it checks role.

Below is an example:

@Get(':house_id')
@UseGuards(HouseRolesGuard)
@HouseRoles(HouseRole.TENANT, HouseRole.ADMIN)